+++ packages/manager/.changeset/pr-13107-tests-1763492424120.md
@@ -1,0 +1,5 @@
+---
+"@linode/manager": Tests
+---
+
+Fixed various test failures when running tests against Prod environment ([#13107](https://github.com/linode/manager/pull/13107))
