+++ packages/manager/cypress/support/util/cleanup.ts
@@ -15,6 +15,7 @@ import { deleteAllTestSSHKeys } from 'support/api/profile';
 import { deleteAllTestStackScripts } from 'support/api/stackscripts';
 import { deleteAllTestTags } from 'support/api/tags';
 import { deleteAllTestVolumes } from 'support/api/volumes';
+import { enhanceError } from 'support/util/api';
 
 /** Types of resources that can be cleaned up. */
 export type CleanUpResource =
@@ -75,7 +76,20 @@ export const cleanUp = (resources: CleanUpResource | CleanUpResource[]) => {
     for (const resource of resourcesArray) {
       const cleanFunction = cleanUpMap[resource];
       // Perform clean-up sequentially to avoid API rate limiting.
-      await cleanFunction();
+      try {
+        await cleanFunction();
+      } catch (e: any) {
+        // Log a warning but otherwise swallow errors if any resources fail to
+        // be cleaned up. There are a few cases where this is especially helpful:
+        //
+        // - Unplanned API issues or outages resulting in 5xx errors
+        // - 400 errors when inadevertently attempting to delete resources that are still busy (e.g. cleaning up a Linode that is the target of a clone operation)
+        const enhancedError = enhanceError(e);
+        console.warn(
+          'An API error occurred while attempting to clean up one or more resources:'
+        );
+        console.warn(enhancedError.message);
+      }
     }
   };
   return cy.defer(
